feat(asm)!: Mole 0.2 Rust encoder rewrite
Full-tree rewrite of mole-asm/ to produce Mole 0.2 bytecode per
docs/MOLE-0.2-SPEC.md. Single big-bang clean-break commit: the v0
16-bit encoder is deleted, not migrated.

Wire-format changes (pre-Phase-0; AGENTS.md §3.17 permits the
break):

- Instruction width: 16 bits -> 32 bits. Opcode/group/sub stack
  is {group[31:30], sub[29:26]}; flag triple at [2:0] on the
  seven bearers (EMIT_BIT_IMM, EMIT_BIT_REG, EMIT_QUARTER_IMM,
  EMIT_QUARTER_REG, EMIT_BYTE, SAMPLE_BIT_ON_SCL,
  DRIVE_BIT_ON_SCL); reserved-must-be-zero [2:0] on every other
  opcode, including STRETCH_SCL_IMM / STRETCH_SCL_REG.
- Live opcode count: 15 -> 25, across four groups (WIRE=9,
  CTRL=8, DATA=8, LOOP=0-with-16-reserved trap slot).
- BUS_MODE wire values renumbered to {i2c=0, i3c-OD=1,
  i3c-PP=2, hdr-ddr=3} (v0 used {0, 1, 6, 7}).
- BRANCH_ON / WAIT_ON share a 4-bit cond at [16:13] with 10-bit
  signed-offset / unsigned-timeout at [12:3] (was 4-bit / 7-bit
  in v0).
- 2-word bytecode preamble: word 0 = magic 0x4D4C ("ML" LE) +
  version 0x0002; word 1 = body-only length in 32-bit words.
  Preamble is included in the assemble() return value.
- New opcodes: DEC, MOV, ADD_IMM, AND_IMM, OR_IMM, XOR_IMM,
  SHIFT (DATA group); FLAG_CLEAR, MARK, LOAD_TIMING (CTRL
  group expansion); split IMM/REG variants for EMIT_BIT,
  EMIT_QUARTER, STRETCH_SCL; EMIT_BYTE (atomic 8-bit) with
  ACK-pairing rule.
- Retired opcodes: JMP (assembler sugar for BRANCH_ON ALWAYS),
  LOAD_LOOP (assembler sugar for LOAD_IMM R6), DEC_BRANCH
  (split into DEC + BRANCH_ON REG_ZERO).

API changes:

- assemble(...) returns Vec<u32> (was Vec<u16>).
- assemble_to_frame(...) unchanged in signature; frame length
  field now counts 32-bit words; payload is 4 bytes per word
  little-endian; CRC-16/XMODEM unchanged.
- New error variants per docs §13 error catalogue
  (E-LEX-*, E-SYN-*, E-OP-*, E-RANGE-*, E-LBL-*, E-WIRE-*,
  E-PROG-*); each carries its short code for CLI rendering.
- moleasm grammar: mnemonics and cond codes case-insensitive;
  short forms `dom`/`rec` accepted; `(use-raw-primitives)`
  pragma enables five carve-outs (reserved tx_symbol,
  capture-to-non-R7, DRIVE_BIT_ON_SCL dst override, EMIT_BYTE
  unpaired, SET_BUS_MODE numeric literals); `.dw <word>`
  emits a raw 32-bit word under raw-primitives.
- EMIT_BYTE atomicity rule (§13 E-WIRE-003): every mask=1
  EMIT_BYTE must be followed textually by BRANCH_ON MISMATCH
  or FLAG_CLEAR with bit 0 set; mask=0 is exempt.

Test surface:

- 90 unit tests in src/lib.rs (encoder + frame + symbols),
  including one named spec_5_N golden per §12.4 worked
  example (all 25 hand-verified hex values).
- 101 integration tests in tests/adversarial.rs covering the
  §13 error catalogue, raw-primitives carve-outs, BUS_MODE
  renumbering, JMP/LOAD_LOOP sugars, EMIT_BYTE pairing
  (including case-insensitive MISMATCH and equated FLAG_CLEAR
  masks), reserved-name shadowing, register parser strictness.
- 1 doctest in lib.rs.
- 1 ignored stub in tests/golden.rs; v0 fixture-comparison
  golden tests deferred to Phase B4 (which regenerates the
  fixtures themselves).

mole-abi v0.2 constants (PREAMBLE_WORDS, MAX_PROGRAM_WORDS,
MAGIC, etc.) are hardcoded with `// FIXME(B5):` markers in
frame.rs; Phase B5 will move them to mole-abi/ and remove the
literals here.

Outside-scope deferrals (will need separate commits):

- mole-asm-cli (Phase B2) and mole-loader-cli (Phase B6)
  consume the public mole-asm API and break against this
  rewrite. cargo test -p <those crates> will fail until
  those phases land. cargo test -p mole-asm is fully green.
- LINT-001 (REG_ZERO_FLAG overwrite advisory between DEC and
  BRANCH_ON) is a spec-defined warning, not error; deferred
  as a future feat(asm) once the fixture regression suite
  exists to validate it.

BREAKING CHANGE: bytecode wire format bumped to v0.2. The
new format is bit-incompatible with v0 at every level:
instruction width (16 -> 32), opcode layout, BUS_MODE wire
values, BRANCH_ON / WAIT_ON encoding, and the introduction
of a mandatory 2-word preamble. A v0 program loaded into a
v0.2 engine (or vice versa) will mis-decode every word. The
assemble() return type changes from Vec<u16> to Vec<u32>;
downstream callers must update accordingly. Pre-Phase-0 the
wire format is not stable (AGENTS.md §3.17); the encoder
output becomes a stable contract on the first tagged release
of Phase B.

+++ mole-asm/src/lib.rs
@@ -1,33 +1,35 @@
-//! Layer-1 bytecode compiler for the Mole bit-cycle engine.
+//! Layer-1 bytecode compiler for the Mole v0.2 bit-cycle engine.
 //!
-//! `mole-asm` turns moleasm source (the locked assembly grammar
-//! documented in `ROADMAP.md` § "moleasm conventions") into the
-//! 16-bit-word bytecode the FPGA engine executes. It is the Rust port
-//! of the reference Python at `tests/fixtures/mole-asm.py`, which
-//! remains the canonical wire-format spec.
+//! `mole-asm` turns moleasm source (grammar documented in
+//! `docs/MOLE-0.2-SPEC.md` §12) into 32-bit-word bytecode the FPGA
+//! engine executes.  The returned program vector includes a 2-word
+//! preamble (magic + length field).
 //!
 //! # Quick start
 //!
 //! ```
 //! use mole_asm::{assemble, assemble_to_frame};
 //!
-//! // Two equivalent ways to get bytecode out:
+//! // Assemble a minimal program.  The returned vector includes the
+//! // 2-word preamble plus the single HALT instruction.
 //! let words = assemble("HALT status=0\n", "<inline>").unwrap();
-//! assert_eq!(words, vec![0x0000]);
+//! assert_eq!(words.len(), 3); // preamble(2) + body(1)
+//! assert_eq!(words[0], 0x0002_4D4C); // magic + version
+//! assert_eq!(words[1], 1);            // body length = 1 word
+//! assert_eq!(words[2], 0x4000_0000); // HALT status=0
 //!
-//! // Or wrap the bytecode in the host->mole UART frame in one call:
+//! // Or wrap the bytecode in the host->mole UART frame in one call.
+//! // Frame = len(2) + preamble(8) + body(4) + crc(2) = 16 bytes.
 //! let frame = assemble_to_frame("HALT status=0\n", "<inline>").unwrap();
-//! assert_eq!(frame.len(), 6); // len(2) + 1 word(2) + crc(2)
+//! assert_eq!(frame.len(), 16);
 //! ```
 //!
 //! # API stability
 //!
-//! Pre-Phase-0 the wire format is not stabilized (see
-//! `AGENTS.md` §3.17). Once Phase 0 ships, the bytecode produced by
-//! [`assemble`] becomes a stable contract between the host compiler
-//! and every deployed Mole; from that point on, byte-for-byte
-//! reproducibility is enforced by the goldens under
-//! `mole-asm/tests/fixtures/`.
+//! Pre-Phase-0 the wire format is not stabilised (see `AGENTS.md`
+//! §3.17 and `docs/MOLE-0.2-SPEC.md` §1).  Once Phase 0 ships, the
+//! bytecode produced by [`assemble`] becomes a stable contract between
+//! the host compiler and every deployed Mole.
 
 #![deny(missing_docs)]
 #![warn(clippy::all)]
@@ -40,9 +42,13 @@ mod symbols;
 
 pub use error::{AsmError, Kind, Result, SourceLocation};
 
-/// Assemble moleasm source text into a vector of 16-bit bytecode words.
+/// Assemble moleasm source text into a vector of 32-bit bytecode words.
 ///
-/// `filename` is used only for diagnostics --- it never opens a file.
+/// The returned vector includes the 2-word preamble (word 0 = magic
+/// `0x0002_4D4C`, word 1 = body length in 32-bit words) followed by
+/// the encoded instruction stream.
+///
+/// `filename` is used only for diagnostics — it never opens a file.
 /// Pass a real path when assembling from disk, or a sentinel like
 /// `"<inline>"` / `"<stdin>"` for in-memory sources.
 ///
@@ -52,31 +58,37 @@ pub use error::{AsmError, Kind, Result, SourceLocation};
 /// operand problem the input contains. The error carries a
 /// [`SourceLocation`] so callers can render diagnostics in their own
 /// preferred format.
-pub fn assemble(source: &str, filename: &str) -> Result<Vec<u16>> {
+pub fn assemble(source: &str, filename: &str) -> Result<Vec<u32>> {
     assembler::assemble(source, filename)
 }
 
 /// Assemble moleasm source and wrap the resulting bytecode in the
-/// host-to-Mole UART frame format documented in
-/// `ROADMAP.md` § "Wire format".
+/// host-to-Mole UART frame format.
+///
+/// The frame layout is:
+/// ```text
+/// len_lo, len_hi       u16 LE: total word count (preamble + body)
+/// word_0 .. word_N     32-bit words, each LE (preamble first)
+/// crc_lo, crc_hi       CRC-16/XMODEM over all preceding bytes
+/// ```
 ///
 /// # Errors
 ///
 /// Returns the same [`AsmError`] variants as [`assemble`], plus
-/// [`AsmError::FrameTooLarge`] if the program exceeds the 2048-word
-/// program-memory budget.
+/// [`AsmError::FrameTooLarge`] if the total program size (preamble +
+/// body) exceeds 8194 words.
 pub fn assemble_to_frame(source: &str, filename: &str) -> Result<Vec<u8>> {
     let words = assemble(source, filename)?;
     frame::build_frame(&words)
 }
 
-/// CRC-16/XMODEM helpers used by the framing layer. Re-exported so
-/// downstream tooling (host runtime, on-target sanity checks) can
+/// CRC-16/XMODEM helpers. Re-exported so downstream tooling can
 /// verify frames without depending on the framing module directly.
 pub mod crc {
     /// Compute the CRC-16/XMODEM of `bytes` (polynomial `0x1021`,
-    /// init `0x0000`, no reflection, no final XOR). Catalog check
-    /// value: `crc(b"123456789") == 0x31C3`.
+    /// init `0x0000`, no reflection, no final XOR).
+    ///
+    /// Catalog check value: `crc(b"123456789") == 0x31C3`.
     pub fn xmodem(bytes: &[u8]) -> u16 {
         crate::frame::crc16_xmodem(bytes)
     }
